+++ packages/runed/src/lib/functions/index.ts
@@ -1,10 +1,10 @@
 export * from "./box/index.js";
 export * from "./activeElement/index.js";
 export * from "./useDebounce/index.js";
-export * from "./useElementSize/index.js";
+export * from "./ElementSize/index.js";
 export * from "./useEventListener/index.js";
 export * from "./useMounted/index.js";
 export * from "./useSupported/index.js";
-export * from "./useStateHistory/index.js";
-export * from "./usePrevious/index.js";
+export * from "./StateHistory/index.js";
+export * from "./Previous/index.js";
 export * from "./watch/index.js";
