--- packages/runed/src/lib/functions/useElementSize/index.ts
@@ -1,1 +1,0 @@
-export * from "./useElementSize.svelte.js";
