--- packages/runed/src/lib/functions/usePrevious/index.ts
@@ -1,1 +1,0 @@
-export * from "./usePrevious.svelte.js";
