--- packages/runed/src/lib/functions/useStateHistory/index.ts
@@ -1,1 +1,0 @@
-export * from "./useStateHistory.svelte.js";
